GARBLED PACMAN

+++ src/main.cpp
@@ -135,7 +135,7 @@ void NESLoop(ROM* r_ptr) {
         }
         if (ppu.debug) {
             printf("PPU REGISTERS: ");
-            printf("VBLANK: %i, PPUCTRL: %02x, PPUMASK: %02x, PPUSTATUS: %02x, OAMADDR: N/A (so far), PPUADDR: %04x\n",ppu.vblank, cpu.memory[0x2000],cpu.memory[0x2001],cpu.memory[0x2002],ppu.v);
+            printf("VBLANK: %i, PPUCTRL: %02x, PPUMASK: %02x, PPUSTATUS: %02x, OAMADDR: N/A (so far), PPUADDR: %04x\n",ppu.vblank, (uint8_t)cpu.memory[0x2000],(uint8_t)cpu.memory[0x2001],(uint8_t)cpu.memory[0x2002],ppu.v);
             printf("scanline: %i, cycle: %i\n",ppu.scanline,ppu.scycle);
         }
         total_ticks = cpu.cycles;
